+++ Apps/phdigitalshadows/dsapi/model/intelligence_incident.py
@@ -47,6 +47,6 @@ class IntelligenceIncident(DSModel):
 
     @classmethod
     def from_json(cls, json):
-        incident_id = DSModel.cast(json.get('id'), long)
+        incident_id = DSModel.cast(json.get('id'), int)
         return cls(incident_id, json)
